This adds SASS files
Update of the sass/css, I checked and the version of nomralize is
the same as the one in the original. The only difference is the
last line that points to the source map. Also, this might be a
linux only thing, but the capitalization on the imports of the
UI folder had to be matched in case.